Implementation of the interference and total Function
I add the Functions in Mfinite_HardPart_2.cpp and .h files to compute
the interference between top and bottom and the total contribution. I
implement both the functions and their derivatives. In order to simplify
implementation of the interference and of total corrections I add to F
the factor yt^2 or yb^2 or yt yb that first was added in the functions
hpt_finite and C of file RHEHpt.cpp. Now I am going to change RHEHpt.cpp
adding a choice number (1,2,3,4) to select the four possible
computations (top^2 term bottom^2 term interference term or total term).
Maybe a switch....