Implementation of the various contribution to top-bottom case
I add to the class RHEHpt two private variable tha mass of the bottom
and a int choice to select one of the possible four calculation (1 only top,
2 only bottom, 3 bottom-top interference,4 total).
I also implement all the core functions to perform the different integration
in the four cases. Up to now the selection is performed by a switch in
functions I_1_2 and I_3 for the coefficients, hpt_finite
for the resummed result. Another switch is inserted in the C function to
select the right form for the LO coefficient.
With these corrections, the program is supposed to evaluate correctly all
the four contribution to the general coefficient c_j,k and to the resummed
result.