Implmented interface with LHAPDF6
Now RHEHpt constructor creates a pointer to the central member
of the PDFSet chosen and uses it to evaluate the coupling constant
at the scale mH.

Removed the Makefile from the .gitignore file and updated it
in order to compile with the lhapdf flags and libs.
Set CUBA_DIR in .bashrc to compile with the new make.

Removed old files "supporting_functions" and added Inverse_Mellin
to store the implementation of the Inverse Mellin Transform