Improved choice between top,bottom,interference and total
Now the choice is made only at the level of the C(i,j).
No more multiplied functions.
Created a new class _parameters to be passed to the Cuba integrand.
It contains std::function data member in order to tell the integrand
the functions to use in Mfinite_HardPart.

Add .gitconfig to .gitignore