Correction of DCO divergent behaviour
I insert an if else in the DC0 function in order to prevent the
indeterminate form 0/0 when x1 or x2 approach zero. I checked that
the small xp limit is obtained for all the part in the integration error.
Moreover I insert a ? construction in the DF0 and related function which
substitute DF0 with the right form at small xp. In this way we fast the program
since we avoid many computation which brings negiglible contributions,
and we can reach value of xp down to 10^-15. Without this correction under
10^-12 the integration fails to converge.