Implementation of exact orders up to NLO in the pointlike approximation.
The class NLOPL permits to compute the LO and NLO of the pointlike
transverse momentum distribution following results of Glosser Schmidt
http://jhep.sissa.it/archive/papers/jhep122002016/jhep122002016.pdf.
The program integrates the double differential cross section over rapidity
and writes the result dsigma/dxp as a function of xp=pt^2/mh^2 and x=mh^2/CME.
The as value at the mass of the Higgs is taken from a PDF that up to know
is of no other use.

I'm going to implement convolution with PDF and to add to this class a flag to
hadronic evaluation.